refactor: extract PacksFacade from CLIFacade and delete monolith
Move pack management methods (listPacks, showPack, installPack,
uninstallPack, configurePack, resolvePack, makePackManager) and all
pack CLI types to standalone PacksFacade. CLIFacade.swift now contains
only CLIVersion and printErr. Rename CLIFacadeTests.swift to
CollectionsFacadeTests.swift to match its class name.

Closes #443
