+++ Sources/KeyPathCLI/Commands/CompletionsCommand.swift
@@ -216,7 +216,7 @@ struct Completions: ParsableCommand {
         mutating func run() async throws {
             switch noun.lowercased() {
             case "pack":
-                let facade = await MainActor.run { CLIFacade() }
+                let facade = PacksFacade()
                 let packs = await facade.listPacks()
                 for pack in packs {
                     let slug = pack.id.replacingOccurrences(of: "com.keypath.pack.", with: "")
