+++ Sources/KeyPathCLI/Commands/Pack/PackUninstallCommand.swift
@@ -18,7 +18,7 @@ struct PackUninstall: AsyncParsableCommand {
 
     mutating func run() async throws {
         let ctx = globals.outputContext
-        let facade = await MainActor.run { CLIFacade() }
+        let facade = PacksFacade()
 
         let spinner = CLISpinner(context: ctx)
